Move scarb related code to scarb.rs
